+++ frontend/packages/akr/src/interfaces/clerkTranslatorTextField.ts
@@ -1,11 +1,25 @@
 import { ChangeEvent } from 'react';
 import { CustomTextFieldProps } from 'shared/components';
 
-import { ClerkTranslatorBasicInformation } from 'interfaces/clerkTranslator';
+import {
+  ClerkTranslatorAddressFieldEnum,
+  ClerkTranslatorTextFieldEnum,
+} from 'enums/clerkTranslator';
+import {
+  ClerkTranslatorAddress,
+  ClerkTranslatorBasicInformation,
+} from 'interfaces/clerkTranslator';
 
-export type ClerkTranslatorTextFieldProps<T> = {
+export type ClerkTranslatorAddressFieldProps = {
+  translator?: ClerkTranslatorAddress;
+  field: ClerkTranslatorAddressFieldEnum;
+  showFieldError: boolean;
+  onChange: (e: ChangeEvent<HTMLTextAreaElement>) => void;
+} & CustomTextFieldProps;
+
+export type ClerkTranslatorTextFieldProps = {
   translator?: ClerkTranslatorBasicInformation;
-  field: T;
+  field: ClerkTranslatorTextFieldEnum;
   showFieldError: boolean;
   onChange: (e: ChangeEvent<HTMLTextAreaElement>) => void;
 } & CustomTextFieldProps;
